new foulder and BLUE detection kinda works
Made a new foulder for the images sorted by hash. Also, added to the hahs sorting a function that detects certain shades of blue in the images, with the hopes that this will detect the irridesecent near the edges of couds. Then after detecting the blue for each images add a t/f var to the renamed image to see it easier. Results...if no blue is detected then there are no couds in the image (this is 100% correct!!!!). However, there are still a lot of false positives where there are no couds but it detects blue anyway....maybe due to the snow.
-MSL